requirements.txt
livekit==1.0.6
livekit-api~=1.0.4

livekit-agents>=1.0.0,<2.0.0
livekit-plugins-openai>=1.0.0,<2.0.0
livekit-plugins-deepgram>=1.0.0,<2.0.0
livekit-plugins-cartesia>=1.0.0,<2.0.0
livekit-plugins-silero>=1.0.0,<2.0.0
livekit-plugins-noise-cancellation==0.2.5
livekit-plugins-turn-detector>=1.0.0,<2.0.0

python-dotenv
fastapi==0.111.0
uvicorn[standard]==0.29.0
openai>=1.0.0
httpx


